Add `setProjectionsToCamera`/`Screen` methods
Utilities to align the rendering objects (`BATCH`, `SHAPES`) to the projection matrix of the world camera / the screen. Useful for switching between world rendering and GUI rendering. This version currently stretches when using the screen projection matrix, because it simply resets to its initial value. We need something smarter to achieve a centered GUI effect.